Omp parallel compiles in VS only with int.

+++ hydra_api/HRExtensions_Spectral.cpp
@@ -282,7 +282,7 @@ namespace hr_spectral
     std::vector<float> imgData(a_width * a_height * channels);
 
     #pragma omp parallel for
-    for(size_t i = 0; i < a_width * a_height; ++i)
+    for(int i = 0; i < a_width * a_height; ++i)
     {
       std::vector<float> spectralPixel(spec.size());
       for(size_t j = 0; j < spec.size(); ++j) // TODO: bad memory access, worth it to fix?
@@ -343,7 +343,7 @@ namespace hr_spectral
     std::vector<float> imgData(a_width * a_height);
 
 #pragma omp parallel for
-    for(size_t i = 0; i < a_width * a_height; ++i)
+    for(int i = 0; i < a_width * a_height; ++i)
     {
       std::vector<float> spectralPixel(spec.size());
       for(size_t j = 0; j < spec.size(); ++j) // TODO: bad memory access, worth it to fix?
@@ -361,7 +361,7 @@ namespace hr_spectral
     int dataChannels = (channels == 4) ? 3 : channels; // alpha channel does not store spectral data
     // add all channels from buf to imgData
     #pragma omp parallel for
-    for(size_t i = 0; i < imgData.size(); ++i)
+    for(int i = 0; i < imgData.size(); ++i)
     {
       for(int k = 0; k < dataChannels; ++k)
       {
@@ -382,7 +382,7 @@ namespace hr_spectral
       {
         std::vector<float> oneChannel(a_width * a_height);
 #pragma omp parallel for
-        for (size_t j = 0; j < a_width * a_height; ++j)
+        for (int j = 0; j < a_width * a_height; ++j)
         {
           oneChannel[j] = buffer[j * channels + k];
         }
@@ -446,7 +446,7 @@ namespace hr_spectral
 
     const float divisor = 1.0f / float(totalSpectralBands);
     #pragma omp parallel for
-    for(size_t j = 0; j < imgData.size(); ++j)
+    for(int j = 0; j < imgData.size(); ++j)
     {
       imgData[j] *= divisor;
     }
@@ -464,7 +464,7 @@ namespace hr_spectral
 
     std::vector<float> imgData(width * height);
     #pragma omp parallel for
-    for(size_t i = 0; i < width * height; ++i)
+    for(int i = 0; i < width * height; ++i)
     {
       std::vector<float> spectralPixel(spectralData.size());
       for(size_t j = 0; j < spectralData.size(); ++j) // TODO: bad memory access, worth it to fix?
